restapi/requirements.txt
Django==4.2.13
django-debug-toolbar==4.3.0
django-filter==24.2
psycopg2-binary==2.9.9
djangorestframework==3.15.1
requests==2.31.0
django-cors-headers==4.3.1
drf-spectacular[sidecar]==0.27.2
django_redis==5.4.0
JSON-log-formatter==1.0
whitenoise==6.6.0
sentry-sdk==2.1.1
pygraphviz==1.12
django-prometheus==2.3.1
gunicorn[gevent]==22.0.0
pythonping==1.1.4
psycogreen==1.0.2
django-db-geventpool==4.0.2
django-select2==8.1.2
ipython==8.24.0